requirements.txt
revolt.py
python-dotenv
